+++ xugrid/core/dataset_accessor.py
@@ -232,6 +232,7 @@ class UgridDatasetAccessor(AbstractUgridAccessor):
         Returns
         -------
         points: Union[xr.DataArray, xr.Dataset]
+            The name of the topology is prefixed in the x, y coordinates.
         """
         result = self.obj
         for grid in self.grids:
@@ -280,6 +281,52 @@ class UgridDatasetAccessor(AbstractUgridAccessor):
             xx, yy, index = grid.rasterize_like(x, y)
             datasets.append(self._raster(xx, yy, index))
         return xr.merge(datasets)
+
+    def intersect_line(
+        self, start: Sequence[float], end: Sequence[float]
+    ) -> xr.Dataset:
+        """
+        Intersect a line with the grid of this data, and fetch the values of
+        the intersected faces.
+
+        Parameters
+        ----------
+        obj: xr.DataArray or xr.Dataset
+        start: sequence of two floats
+            coordinate pair (x, y), designating the start point of the line.
+        end: sequence of two floats
+            coordinate pair (x, y), designating the end point of the line.
+
+        Returns
+        -------
+        intersection: xr.Dataset
+            The name of the topology is prefixed in the x, y and s
+            (spatium=distance) coordinates.
+        """
+        obj = self.obj
+        for grid in self.grids:
+            obj = grid.intersect_line(obj, start, end)
+        return obj
+
+    def intersect_linestring(self, linestring) -> xr.Dataset:
+        """
+        Intersect the grid along a collection of linestrings. Returns a new Dataset
+        with the values for each intersected segment.
+
+        Parameters
+        ----------
+        linestring: shapely.LineString
+
+        Returns
+        -------
+        intersection: xr.Dataset
+            The name of the topology is prefixed in the x, y and s
+            (spatium=distance) coordinates.
+        """
+        obj = self.obj
+        for grid in self.grids:
+            obj = grid.intersect_linestring(obj, linestring)
+        return obj
 
     def to_dataset(self, optional_attributes: bool = False):
         """