+++ docs/changelog.rst
@@ -9,13 +9,31 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 Unreleased
 ----------
 
+Added
+~~~~~
+
+- :meth:`xugrid.Ugrid2d.intersect_line`,
+  :meth:`xugrid.Ugrid2d.intersect_linestring`
+  :meth:`xugrid.UgridDataArrayAccessor.intersect_line`,
+  :meth:`xugrid.UgridDataArrayAccessor.intersect_linestring`,
+  :meth:`xugrid.UgridDatasetAccessor.intersect_line`, and
+  :meth:`xugrid.UgridDatasetAccessor.intersect_linestring` have been added to
+  intersect line and linestrings and extract the associated face data.
+
+Changed
+~~~~~~~
+
+- Selection operations along a line, or at point locations, will now prefix the
+  name of the grid in the x and y coordinates. This avoids name collisions when
+  multiple topologies are present in a dataset.
+
 [0.6.4] 2023-08-22
 ------------------
 
 Fixed
 ~~~~~
 
-- Bug in :func:`xugrid.snap_to_grid`, which caused an ``IndexError``. 
+- Bug in :func:`xugrid.snap_to_grid`, which caused an ``IndexError``.
   See `#122 <https://github.com/Deltares/xugrid/issues/122>`_.